requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.3
django-filter==22.1
django-health-check==3.17.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
djangorestframework-simplejwt==5.3.1
environs==9.5.0
psycopg2==2.9.5
pytest==7.2.1
djangorestframework-api-key==2.*
flake8==6.0.0
ipython==8.10.0
ipdb==0.13.13
isort==5.12.0
pre-commit==3.0.4
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-env==0.8.1
sort-requirements==1.3.0
PyJWT==1.7.1
